fixed double health and added output file

+++ BattleUnits/AlienDrone.cpp
@@ -5,7 +5,7 @@
 
 
 
-AlienDrone::AlienDrone(Game* game, int joinTime, int health, int power, int attackCapacity)
+AlienDrone::AlienDrone(Game* game, int joinTime, double health, double power, int attackCapacity)
     : Unit(game,AD,joinTime, health, power, attackCapacity) {
 }
