+++ BattleUnits/AlienMonster.cpp
@@ -4,7 +4,7 @@
 #include "../DataStructures/LinkedListStack.h"
 
 
-AlienMonster::AlienMonster(Game* game, int joinTime, int health, int power, int attackCapacity)
+AlienMonster::AlienMonster(Game* game, int joinTime, double health, double power, int attackCapacity)
     : Unit(game,AM, joinTime, health, power, attackCapacity) {
 }
