+++ BattleUnits/AlienMonster.h
@@ -5,7 +5,7 @@
 
 class AlienMonster : public Unit {
    public:
-    AlienMonster(Game* game, int joinTime, int health, int power, int attackCapacity);
+    AlienMonster(Game* game, int joinTime, double health, double power, int attackCapacity);
     void attack()override;
 };
