+++ BattleUnits/EarthGunnery.cpp
@@ -4,7 +4,7 @@
 #include "../DataStructures/LinkedListStack.h"
 
 
-EarthGunnery::EarthGunnery(Game* game, int joinTime, int health, int power, int attackCapacity)
+EarthGunnery::EarthGunnery(Game* game, int joinTime, double health, double power, int attackCapacity)
     : Unit(game,EG, joinTime, health, power, attackCapacity) {
 };
