+++ BattleUnits/EarthGunnery.h
@@ -5,7 +5,7 @@
 
 class EarthGunnery : public Unit {
    public:
-    EarthGunnery(Game* game, int joinTime, int health, int power, int attackCapacity);
+    EarthGunnery(Game* game, int joinTime, double health, double power, int attackCapacity);
     void attack() override;
     int getPriority();
 };
