+++ BattleUnits/EarthTank.h
@@ -7,7 +7,7 @@ class EarthTank : public Unit
 {
 
 public:
-	EarthTank(Game* game, int joinTime, int health, int power, int attackCapacity);
+	EarthTank(Game* game, int joinTime, double health, double power, int attackCapacity);
 	void attack() override;
 };
